requirements.txt
numpy
pandas
matplotlib
scikit-learn
yfinance
torch
gym
ta
tqdm
kiteconnect
python-dotenv